fix: Added multiple sockets instances management in socket-io config.

+++ config/socket-io.config.js
@@ -10,23 +10,46 @@ const initializeSocketIo = (httpServer, CorsOrigin) => {
     }
   });
   io.on('connection', function(socket) {
-    socket.on('setSocketId', async ({userId, socketId}) => {
-      let socketIoDb = await SocketIoModel.findOne({userId : userId});
-      if(socketIoDb){
-        await SocketIoModel.findByIdAndUpdate(socketIoDb._id, { socketId: socketId }, { override: true, upsert: true, new: true });
-      }else{
-        const newSocketIoDb = new SocketIoModel({
-          userId : userId,
-          socketId: socketId
-        });
-        await newSocketIoDb.save();
+    socket.on('setSocketId', async ({userId, socketId, oldSocketId}) => {
+      try {
+        let socketIoDb = await SocketIoModel.findOne({userId : userId});
+        if(socketIoDb){
+          let arraySocketId = socketIoDb.socketId;
+          const indexSocketId = arraySocketId.findIndex(socketId => socketId === oldSocketId);
+          if(indexSocketId !== -1){
+            arraySocketId[indexSocketId] = socketId;
+          }else{
+            arraySocketId.push(socketId);
+          }
+          await SocketIoModel.findByIdAndUpdate(socketIoDb._id, { socketId: arraySocketId }, { override: true, upsert: true, new: true });
+        }else{
+          const newSocketIoDb = new SocketIoModel({
+            userId : userId,
+            socketId: [socketId]
+          });
+          await newSocketIoDb.save();
+        }
+      } catch (error) {
+        console.log(error);
       }
     });
   
     socket.on('disconnect', async () => {
-      await SocketIoModel.findOneAndDelete({
-        socketId : socket.id,
-      });
+      try {
+        let socketIoDb = await SocketIoModel.findOne({socketId : socket.id});
+        let arraySocketId = socketIoDb.socketId;
+        if(arraySocketId.length > 1){
+          const indexSocketId = arraySocketId.findIndex(socketId => socketId === socket.id); 
+          arraySocketId.splice(indexSocketId, 1);
+          await SocketIoModel.findByIdAndUpdate(socketIoDb._id, { socketId: arraySocketId }, { override: true, upsert: true, new: true });
+        }else{
+          await SocketIoModel.findOneAndDelete({
+            socketId : socket.id,
+          });
+        }
+      } catch (error) {
+        console.log(error);
+      }
     });
   });
 }
